Acquire a looper for ExtPhoneCallbackListener while registering
When the unregistered listener is registered back again, it can't
handle any event because the local looper is set as null by the
un-register.

To fix it, need to acquire a proper looper whenever looper is reset
or an external null looper is passed during registering. Besides,
with this change, the initialization is delayed at the point where
receiving a registration, and a extended handler holding a weak
reference of ExtPhoneCallbackListener is used to avoid any possibility
of memory leaks.

Change-Id: Ib57a5bc72ab053681bf6b224e7d62e282760531c
CRs-Fixed: 3986730